pyproject.toml
[project]
name = "acm-hackathon-project-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=5.2.6",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.8.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
]
